Squashed 'modules/rmf/dependency/RMF/' changes from 7671f77..969e42d
969e42d Use 64-bit quantities for block offsets.
8b1d784 Squashed 'tools/dev_tools/' changes from 976d76c..e357d50
1e15f97 Use new method of handling custom command byproducts.

git-subtree-dir: modules/rmf/dependency/RMF
git-subtree-split: 969e42dfc3b2b3c7cf051d4ec552bde4854fec13

+++ modules/rmf/dependency/RMF/src/backend/avro/types.h
@@ -136,7 +136,7 @@ struct FileData {
   RMF_LARGE_UNORDERED_MAP<FrameID, internal::FrameData> frames;
   KeyData keys;
   DataTypes data;
-  RMF_LARGE_UNORDERED_MAP<FrameID, int32_t> frame_block_offsets;
+  RMF_LARGE_UNORDERED_MAP<FrameID, int64_t> frame_block_offsets;
 };
 
 // awkward workaround for Mac OS < 10.9
